backend/requirements.txt
fastapi
pydantic
python-dotenv
pymongo
cryptography
passlib[bcrypt]
python-jose
uvicorn